[INPLACE-169] 마커 관리 hook으로 분리
